+++ tests/testthat/test_frontend.R
@@ -19,6 +19,7 @@ test_that("Frontend works remotely", {
   expect_type(data_vintage(id), "character")
   expect_type(data_metadata(id), "list")
   expect_s3_class(data_search(id), "robonomist_search")
+  expect_s3_class(data(), "robonomist_search")
   expect_s3_class(datasources(), "robonomist_datasources")
   expect_s3_class(data_get(id), "robonomist_data")
   expect_s3_class(data(id), "robonomist_data")